apps/ocr/requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
python-multipart>=0.0.9
Pillow>=10.3.0
pytesseract>=0.3.10
easyocr>=1.7.1
pdf2image>=1.17.0
numpy>=1.26.0
pydantic>=2.7.0
pydantic-settings>=2.2.0
python-dateutil>=2.9.0
pytest>=8.1.0
pytest-asyncio>=0.23.0
httpx>=0.27.0
